refactor(core): improve directory structure
