+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export { css } from "./css";
 export { createTheme, type Theme, type ThemeSystem } from "./theme";
 export * from "./components";
 export { componentList } from "./components/componentList";
+export * from "./registry";
