+++ packages/core/src/registry/StyleRegistry.tsx
