+++ packages/core/src/registry/index.ts
@@ -1,0 +1,1 @@
+export { StyleRegistry, createStyleRegistry } from "./StyleRegistry";
