+++ packages/core/src/registry/sheet/index.ts
@@ -1,0 +1,1 @@
+export * from "./StyleSheet";
